% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovperm.R
\name{aovperm}
\alias{aovperm}
\title{Integration of the quasi-F statistics into the \code{aovperm} function from the \code{permuco} package.}
\usage{
aovperm(formula, data = NULL, np = 5000, method = NULL,
  effect = NULL, ...)
}
\arguments{
\item{formula}{A formula object. The formula for the quasi-F statistic should be written using the notation
\code{+ Error(id/(A*C)) + Error(id/(B*C))} to specify random effect associate to subjects and items, where \code{A} is a within \code{id} factor, \code{B} is a within \code{item} factor and \code{C} is within \code{id} and \code{item} factors.}

\item{data}{similar to the \code{permuco} package.}

\item{np}{similar to the \code{permuco} package. Default value is \code{5000}.}

\item{method}{similar to the \code{permuco} package. The method is set to \code{"terBraak"} and \code{"terBraak_logp"} are available for the quasi-F statistic.}

\item{effect}{a numeric specifying the effect to test. The value correspond to the "assign" attribute of the model.matrix argument of the fixed effect. The default is NULL tests all effects.}

\item{...}{Futher arguments, see details.}
}
\value{
A \code{lmperm} object containing most of the objects given in an \link{lm} object, an ANOVA table with parametric and permutation p-values, the test statistics and the permutation distributions.
}
\description{
Provides p-values for omnibus tests based on permutations for quasi-F statistics
}
\details{
Similar to the \code{permuco} package.
}
\author{
jaromil.frossard@unige.ch
}
