% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterlm.R
\name{clusterlm}
\alias{clusterlm}
\title{Integration of the quasi-F statistics into the \code{clusterlm} function from the \code{permuco} package.}
\usage{
clusterlm(formula, data = NULL, np = 5000, method = NULL,
  test = "fisher", threshold = NULL, aggr_FUN = NULL,
  multcomp = "clustermass", effect = NULL, ...)
}
\arguments{
\item{formula}{A formula object. The formula for the quasi-F statistic should be written using the notation
\code{+ Error(id/(A*C)) + Error(id/(B*C))} to specify random effect associate to subjects and items, where \code{A} is a within \code{id} factor, \code{B} is a within \code{item} factor and \code{C} is within \code{id} and \code{item} factors.}

\item{data}{similar to the \code{permuco} package.}

\item{np}{The number of permutations. Default value is \code{5000}.}

\item{method}{similar to the \code{permuco} package. The method is set to \code{"terBraak"} and \code{"terBraak_logp"} are available for the quasi-F statistic.}

\item{test}{similar to the \code{permuco} package. the argument does not influence the quasi-F statistic.}

\item{threshold}{similar to the \code{permuco} package.}

\item{aggr_FUN}{similar to the \code{permuco} package.}

\item{multcomp}{similar to the \code{permuco} package.}

\item{effect}{a numeric specifying the effect to test. The value correspond to the "assign" attribute of the model.matrix argument of the fixed effect. The default is NULL tests all effects.}

\item{...}{Futher arguments, see details.}
}
\value{
A list containing : a table of the clusters, or a \code{multcomp} object for the other multiple comparison procedures. Use the \link{plot.clusterlm} method to have a quick overview of the results.
}
\description{
Provides cluster-mass tests based on permutations for quasi-F statistics
}
\details{
Similar to the \code{permuco} package.
}
\examples{

 \dontrun{

# packages

library(permuco)
devtools::install_github("jaromilfrossard/permucoQuasiF")
library(permucoQuasiF)

# data

data("signal")
data("design")

model <- clusterlm( signal ~ A*B*C + Error(id/(B*C)) + Error(item/(A*C)), data=design, np = 5)
plot(model)
model}

}
\author{
jaromil.frossard@unige.ch
}
