% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design_18i_20s}
\alias{design_18i_20s}
\title{Simulated dataset of the design}
\format{A data frame with 720 rows and 5 variables.}
\usage{
data(design_18i_20s)
}
\description{
Simulated dataset of the design of an experiment with 18 items and 20 subject.

\itemize{
\item id : identifier of the subjects.
\item item : identifier of the items.
\item A : between-subject factor with 2 levels.
\item B : between-item factor with 3 levels.
\item B : withen subjects and items factor with 2 levels.
}
}
