% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_z.R
\name{mm_z}
\alias{mm_z}
\title{Create the Z matrices based with sum contrasts}
\usage{
mm_z(model_frame, formula_f, formula_within, formula_id)
}
\arguments{
\item{model_frame}{a model.frame object containint the facotors of interest}

\item{formula_f}{a formula of all the effects as fixed}

\item{formula_within}{a formula indicating the within effect}

\item{formual_id}{the formula indicating the sampling units}
}
\value{
a matrix with its contrasts as attributs.
}
\description{
Compute the z matrice for the random effects and reduced its rank and dimension with respect to the between effects.
}
