% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signal_18i_20s}
\alias{signal_18i_20s}
\title{Simulated dataset of the signal.}
\format{A matrix with 720 rows and 600 variables.}
\usage{
data(signal_18i_20s)
}
\description{
Simulated dataset of the signal of an experiment with 18 items and 20 subject. We assume 600 measures of the signal.
}
